Customize Upstart script block
This changeset switches to an Upstart script block for the Tasseo
Upstart script. This allows users to pass snippets of a script that get
evaluated as the value for `tasseo_graphite_url`.

Example:

```
tasseo_graphite_url: "http://$(curl -s http://instance-data.ec2.internal/latest/meta-data/public-hostname):8080"
```

This would set the Graphite URL to the FQDN of an EC2 instance.